Add files via upload
This repository presents a comparative study on fake and real news detection using state-of-the-art Transformer based models: DistilBERT, BERT, and RoBERTa. The notebook includes data preprocessing, fine tuning of pre trained models, performance evaluation, and analysis. The objective is to leverage Natural Language Processing (NLP) techniques to classify news articles as either real or fake with high accuracy. The project demonstrates the effectiveness of transfer learning in combating misinformation in the digital age.